Add multi-user chat with ActionCable
Things I learned:

* You need to be running a multi-threaded webserver to get capybara
  and actioncable to play nice together, hence the `puma` change
* You also need to allow all the localhost ports as request origins,
  since capybara/poltergeist/someone swaps around all the ports
  and that makes trouble

The JS is starting to inflate a little bit and I'm not too happy having
the channel code just sitting in there, although it's ostensibly
where it belongs at the moment. We'll see where that one goes. Also the
ChatService API needs changed because we're no longer really 'submitting'
messages.

Oh, and our own messages appear twice! Will be adding a test for that.