Handle exceptions in Assembler interpreting
We make `AssemblerInterpreter` throw a single sort of exception, then
create a `TolerantAssemblerInterpreter` which swallows them and returns
the original state.

The idea is that messages might contain invalid Assembler, but will
refuse to execute. Though maybe errors would be nicer in the
long term...

We've also been rejigging a few things to clean up the architecture