Implement `Assembler::State` for storing machine state
It's effectively an immutable hash with a bit of an explicit API to
guide usage. You modify it with `#generate`, which yields to a block
with a hash argument that you can modify freely. It then returns a new
`Assembler::State`. The idea behind the name is that you're creating
a new 'generation' by calling that method.

We'll see if it sticks